drf_web_two/requirements.txt
aliyun-python-sdk-core==2.13.12
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.8.0
asn1crypto==1.2.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==2.8
Django==2.2.5
django-ckeditor==5.8.0
django-cors-headers==3.2.0
django-filter==2.2.0
django-js-asset==1.2.2
django-redis==4.11.0
django-simpleui==3.7
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
mysqlclient==1.3.14
olefile==0.46
oss2==2.9.0
packaging==19.2
Pillow==6.2.1
pycparser==2.19
pycryptodome==3.9.4
pycryptodomex==3.9.4
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.5
PySocks==1.7.1
python-alipay-sdk==2.0.1
pytz==2019.3
redis==3.3.11
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
win-inet-pton==1.1.0
wincertstore==0.2
